chore: explicitly include matrix in tsconfig
