+++ e2e/tsconfig.json
@@ -2,13 +2,13 @@
   "compilerOptions": {
     "esModuleInterop": true,
     "lib": ["ES2022"],
-    "module": "CommonJS",
+    "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "strict": true,
     "target": "ES2022",
     "useUnknownInCatchVariables": false
   },
-  "include": ["./**/*.e2e.test.ts"],
+  "include": ["./**/*.e2e.test.ts", "./matrix.ts"],
   "exclude": ["node_modules", "dist", "**/crds/**"]
 }
